+++ src/image/image_handler.h
@@ -12,8 +12,8 @@ public:
 
   ~image_handler();
 
-  void draw_rotate(const std::string name, double x, double y, double frame_bump, double angle);
-  void draw_tile(const std::string name, double parallax);
+  void draw_rotate(const std::string name, float x, float y, float frame_bump, float angle);
+  void draw_tile(const std::string name, float parallax);
 
 private:
   std::unordered_map<std::string, image> images;
