+++ src/vec2d/vec2d.cpp
@@ -6,7 +6,7 @@
 #include <math.h>
 
 //these operators let us return values with array notation
-double &vec2d::operator [](int i) {
+float &vec2d::operator [](int i) {
   //return a reference
   switch(i) {
     case 0:
@@ -19,7 +19,7 @@ double &vec2d::operator [](int i) {
       return max;
   }
 }
-double vec2d::operator [](int i) const {
+float vec2d::operator [](int i) const {
   //return a const val
   switch(i) {
     case 0:
@@ -35,19 +35,19 @@ double vec2d::operator [](int i) const {
 
 //-----------------------------------------------------------------------------
 //scalar const operations
-vec2d vec2d::operator+(double scalar_val) const {
+vec2d vec2d::operator+(float scalar_val) const {
   return vec2d(x + scalar_val, y + scalar_val);
 }
 
-vec2d vec2d::operator-(double scalar_val) const {
+vec2d vec2d::operator-(float scalar_val) const {
   return vec2d(x - scalar_val, y - scalar_val);
 }
 
-vec2d vec2d::operator*(double scalar_val) const {
+vec2d vec2d::operator*(float scalar_val) const {
   return vec2d(x * scalar_val, y * scalar_val);
 }
 
-vec2d vec2d::operator/(double scalar_val) const {
+vec2d vec2d::operator/(float scalar_val) const {
   if(scalar_val == 0) { throw "can't divide vector by 0!\n"; }
   return vec2d(x / scalar_val, y / scalar_val);
 }
@@ -79,28 +79,28 @@ void vec2d::operator-=(const vec2d &v) {
 //-----------------------------------------------------------------------------
 //related to vector properties
 
-double vec2d::magnitude() const {
+float vec2d::magnitude() const {
   return sqrt((x * x) + (y * y));
 }
 
-double vec2d::magnitudeSquared() const {
+float vec2d::magnitudeSquared() const {
   return ((x * x) + (y * y));
 }
 
 vec2d vec2d::normalize() const {
   //NOTE: will not normalize beyond .001, returns this
-  double m = this->magnitude();
+  float m = this->magnitude();
   if(m < .001 && m > -.001) {
     return *this;
   }
   return vec2d(x / m, y / m);
 }
 
-double vec2d::dot(const vec2d &v) const {
+float vec2d::dot(const vec2d &v) const {
   return ((x * v[0]) + (y * v[1]));
 }
 
-vec2d vec2d::cap(double scalar_val) const {
+vec2d vec2d::cap(float scalar_val) const {
   //"cap" a vector's magnitude at a certain value
 
   if(this->magnitude() > scalar_val) {
@@ -113,9 +113,9 @@ vec2d vec2d::cap(double scalar_val) const {
   }
 }
 
-vec2d vec2d::decay(double scalar_val) const {
+vec2d vec2d::decay(float scalar_val) const {
   //"decay" a vector by removing a fraction of its magnitude
-  double m = this->magnitude();
+  float m = this->magnitude();
   if(m < 0.05 && m > -0.05) {
     return(vec2d(0,0));
   }
@@ -126,11 +126,11 @@ vec2d vec2d::decay(double scalar_val) const {
 //------------------------------------------------------------------------------
 //compute angles from the vectors
 
-double vec2d::angle_rad() const {
+float vec2d::angle_rad() const {
 return(std::atan2(y, x));
 }
 
-double vec2d::angle_deg() const {
+float vec2d::angle_deg() const {
   //return an angle between 0 and 360
   return(int((this->angle_rad() * (180/PI) + 90) + 360) % 360);
 }
